Create unitTest result
This document lists a series of pylint test results. I am very surprised that there are so many non-compliances in a piece of code, although these problems will not cause errors. For example, if there are a total of three cases, then only need: if.. else.. At this time, there is no need to be at home else, because if the first two are not met, then the third case will be run directly. And, pylint can detect unnecessary end-of-sentence spaces, line breaks, new lines at the end, and so on. After testing and modifying, my gains are very big